+++ frontend/src/pages/public/index.tsx
@@ -1,0 +1,1 @@
+export { PublicItemDetailPage } from "./PublicItemDetailPage";
